requirements.txt
asgiref==3.8.1
Django==3.2.20
django-mysql==4.11.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.2.0
packaging==24.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
